estudos-python-udemy/Projeto de análise de dados/requirements.txt
pandas==1.5.3
streamlit==1.23.1
numpy==1.25.2
matplotlib==3.8.0
seaborn==0.12.2
plotly.express